test(workspace-metadata): add comprehensive tests for workspace metadata loading
- Add extensive unit tests for cargo metadata loading, workspace graph building,
  and cmd-mox integration.
- Introduce test fixtures and error scenarios for improved coverage.
- Remove deprecated or redundant workspace_metadata tests to consolidate test suite.
- Tests cover edge cases, error handling, and environment configurations related
  to cargo metadata processing and workspace model construction.
