+++ tests/helpers/workspace_metadata.py
@@ -2,13 +2,24 @@
 
 from __future__ import annotations
 
-__all__ = ["build_test_package", "create_test_manifest"]
+__all__ = ["ErrorScenario", "build_test_package", "create_test_manifest"]
 
+import dataclasses as dc
 import textwrap
 import typing as typ
 
 if typ.TYPE_CHECKING:
     from pathlib import Path
+
+
+@dc.dataclass(frozen=True, slots=True)
+class ErrorScenario:
+    """Test scenario for cargo metadata error cases."""
+
+    exit_code: int
+    stdout: str
+    stderr: str
+    expected_message: str
 
 
 class DependencyEntry(typ.TypedDict, total=False):
